# Add fc padding to improve mkl GEMM's performance when N and K are multiple of 128.