--- a/paddle/fluid/operators/fc_op.cc
+++ b/paddle/fluid/operators/fc_op.cc
@@ -41,9 +41,10 @@ class FCOp : public framework::OperatorWithKernel {
         PADDLE_ENFORCE_EQ(bias_dims[0], 1,
                           platform::errors::InvalidArgument(
                               "The shape of Bias should be [1, dim]."));
-        PADDLE_ENFORCE_EQ(bias_dims[1], w_dims1,
-                          platform::errors::InvalidArgument(
-                              "The width of Bias should be equal to Weight."));
+        PADDLE_ENFORCE_EQ(
+            bias_dims[1], w_dims1,
+            platform::errors::InvalidArgument(
+                "The width of Bias should be equal to width of Weight."));
       } else if (bias_dims.size() == 1) {
         PADDLE_ENFORCE_EQ(
             bias_dims[0], w_dims1,
@@ -116,10 +117,6 @@ class FCOpMaker : public framework::OpProtoAndCheckerMaker {
     AddAttr<bool>("padding_weights",
                   "(bool, default false) Only used when weight padding in pass")
         .SetDefault(false);
-    AddAttr<bool>(
-        "use_gpu",
-        "(bool, default true) Guaranteed weight padding is used in cpu scope")
-        .SetDefault(true);
     AddAttr<bool>(framework::kAllKernelsMustComputeRuntimeShape,
                   "Skip calling InferShape() function in the runtime.")
         .SetDefault(true);