--- a/paddle/fluid/operators/fc_op.cc
+++ b/paddle/fluid/operators/fc_op.cc
@@ -40,16 +40,25 @@ class FCOp : public framework::OperatorWithKernel {
       if (bias_dims.size() == 2) {
         PADDLE_ENFORCE_EQ(bias_dims[0], 1,
                           platform::errors::InvalidArgument(
-                              "The shape of Bias should be [1, dim]."));
+                              "The shape of Bias is invalid."
+                              "The height of Bias should be 1."
+                              "But reveived height of Bias is %d.",
+                              bias_dims[0]));
         PADDLE_ENFORCE_EQ(
             bias_dims[1], w_dims1,
             platform::errors::InvalidArgument(
-                "The width of Bias should be equal to width of Weight."));
+                "The shape of Bias is invalid."
+                "The width of Bias should be equal to width of Weight."
+                "But reveived width of Bias is %d and width of Weight is %d.",
+                bias_dims[1], w_dims1));
       } else if (bias_dims.size() == 1) {
         PADDLE_ENFORCE_EQ(
             bias_dims[0], w_dims1,
             platform::errors::InvalidArgument(
-                "The shape of Bias should be equal to the width of weight."));
+                "The shape of Bias is invalid."
+                "The height of Bias should be equal to the width of weight."
+                "But reveived height of Bias is %d and width of Weight is %d.",
+                bias_dims[0], w_dims1));
       }
     }