--- a/paddle/fluid/operators/fc_op.h
+++ b/paddle/fluid/operators/fc_op.h
@@ -29,25 +29,18 @@ inline void FCOutputSize(const framework::DDim& in_dims,
                          std::vector<int64_t>& out_dims,  // NOLINT
                          int in_num_col_dims, bool weight_pass) {
   auto in_mat_dims = framework::flatten_to_2d(in_dims, in_num_col_dims);
-  if (weight_pass) {
-    PADDLE_ENFORCE_EQ(
-        in_mat_dims[1], w_dims[0] - 4,
-        "Fully Connected input and weigth size do not match. %s, %s");
-  } else {
-    PADDLE_ENFORCE_EQ(
-        in_mat_dims[1], w_dims[0],
-        "Fully Connected input and weigth size do not match. %s, %s");
-  }
+  auto w_dims0 = weight_pass ? w_dims[0] - 4 : w_dims[0];
+  auto w_dims1 = weight_pass ? w_dims[1] - 4 : w_dims[1];
+  PADDLE_ENFORCE_EQ(
+      in_mat_dims[1], w_dims0,
+      platform::errors::InvalidArgument(
+          "Fully Connected input and weigth size do not match. %s, %s"));
 
   out_dims.reserve(static_cast<size_t>(in_num_col_dims + 1));
   for (int i = 0; i < in_num_col_dims; ++i) {
     out_dims.push_back(in_dims[i]);
   }
-  if (weight_pass) {
-    out_dims.push_back(w_dims[1] - 4);
-  } else {
-    out_dims.push_back(w_dims[1]);
-  }
+  out_dims.push_back(w_dims1);
 }
 
 template <typename DeviceContext, typename T>
@@ -72,12 +65,9 @@ class FCOpKernel : public framework::OpKernel<T> {
     output->set_lod(input->lod());
 
     auto out_dims = output->dims();
-    int M = 0;
-    if (weight_pass) {
-      M = framework::product(out_dims) / (w_dims[1] - 4);
-    } else {
-      M = framework::product(out_dims) / w_dims[1];
-    }
+    auto w_dims0 = weight_pass ? w_dims[0] - 4 : w_dims[0];
+    auto w_dims1 = weight_pass ? w_dims[1] - 4 : w_dims[1];
+    int M = framework::product(out_dims) / w_dims1;
 
     const T* input_data = input->data<T>();
     const T* w_data = w->data<T>();
@@ -85,13 +75,8 @@ class FCOpKernel : public framework::OpKernel<T> {
 
     auto& dev_ctx = ctx.template device_context<DeviceContext>();
     math::FCFunctor<DeviceContext, T> fc;
-    if (weight_pass) {
-      fc(dev_ctx, M, w_dims[1] - 4, w_dims[0] - 4, input_data, w_data,
-         output_data, bias ? bias->data<T>() : NULL, with_relu, weight_pass);
-    } else {
-      fc(dev_ctx, M, w_dims[1], w_dims[0], input_data, w_data, output_data,
-         bias ? bias->data<T>() : NULL, with_relu);
-    }
+    fc(dev_ctx, M, w_dims1, w_dims0, input_data, w_data, output_data,
+       bias ? bias->data<T>() : NULL, with_relu, weight_pass);
   }
 };