--- a/paddle/fluid/operators/fc_op.h
+++ b/paddle/fluid/operators/fc_op.h
@@ -27,10 +27,10 @@ using Tensor = framework::Tensor;
 inline void FCOutputSize(const framework::DDim& in_dims,
                          const framework::DDim& w_dims,
                          std::vector<int64_t>& out_dims,  // NOLINT
-                         int in_num_col_dims, bool weight_pass) {
+                         int in_num_col_dims, bool padding_weights) {
   auto in_mat_dims = framework::flatten_to_2d(in_dims, in_num_col_dims);
-  auto w_dims0 = weight_pass ? w_dims[0] - 4 : w_dims[0];
-  auto w_dims1 = weight_pass ? w_dims[1] - 4 : w_dims[1];
+  auto w_dims0 = padding_weights ? w_dims[0] - 4 : w_dims[0];
+  auto w_dims1 = padding_weights ? w_dims[1] - 4 : w_dims[1];
   PADDLE_ENFORCE_EQ(in_mat_dims[1], w_dims0,
                     platform::errors::InvalidArgument(
                         "Fully Connected input and weigth size do not match. "
@@ -57,17 +57,17 @@ class FCOpKernel : public framework::OpKernel<T> {
         (ctx.Attr<std::string>("activation_type") == "relu") ? true : false;
 
     auto w_dims = w->dims();
-    bool weight_pass = ctx.Attr<bool>("padding_weights");
+    bool padding_weights = ctx.Attr<bool>("padding_weights");
 
     std::vector<int64_t> output_dims;
     FCOutputSize(input->dims(), w_dims, output_dims, in_num_col_dims,
-                 weight_pass);
+                 padding_weights);
     output->Resize(framework::make_ddim(output_dims));
     output->set_lod(input->lod());
 
     auto out_dims = output->dims();
-    auto w_dims0 = weight_pass ? w_dims[0] - 4 : w_dims[0];
-    auto w_dims1 = weight_pass ? w_dims[1] - 4 : w_dims[1];
+    auto w_dims0 = padding_weights ? w_dims[0] - 4 : w_dims[0];
+    auto w_dims1 = padding_weights ? w_dims[1] - 4 : w_dims[1];
     int M = framework::product(out_dims) / w_dims1;
 
     const T* input_data = input->data<T>();
@@ -77,7 +77,7 @@ class FCOpKernel : public framework::OpKernel<T> {
     auto& dev_ctx = ctx.template device_context<DeviceContext>();
     math::FCFunctor<DeviceContext, T> fc;
     fc(dev_ctx, M, w_dims1, w_dims0, input_data, w_data, output_data,
-       bias ? bias->data<T>() : NULL, with_relu, weight_pass);
+       bias ? bias->data<T>() : NULL, with_relu, padding_weights);
   }
 };