--- a/paddle/fluid/operators/math/fc.cc
+++ b/paddle/fluid/operators/math/fc.cc
@@ -25,9 +25,28 @@ class FCFunctor<platform::CPUDeviceContext, T> {
  public:
   void operator()(const platform::CPUDeviceContext& context, const int M,
                   const int N, const int K, const T* X, const T* W, T* Y,
-                  const T* B = nullptr, bool relu = false) {
+                  const T* B = nullptr, bool relu = false,
+                  bool weight_pass = false) {
     auto blas = math::GetBlas<platform::CPUDeviceContext, T>(context);
-    blas.MatMul(M, N, K, X, W, Y);
+    framework::Tensor X1, Y1;
+    X1.Resize({M * (K + 4)});
+    T* X1_data = X1.mutable_data<T>(platform::CPUPlace());
+    Y1.Resize({M * (N + 4)});
+    T* Y1_data = Y1.mutable_data<T>(platform::CPUPlace());
+    if (weight_pass) {
+      const int NN = N + 4;
+      const int KK = K + 4;
+#ifdef PADDLE_WITH_MKLML
+#pragma omp parallel for
+#endif
+      for (int i = 0; i < M; i++) {
+        memcpy(X1_data + i * KK, X + i * K, K * sizeof(X[0]));
+      }
+      blas.GEMM(false, false, M, N, K, static_cast<T>(1.0), X1_data, KK, W, NN,
+                static_cast<T>(0.0), Y1_data, NN);
+    } else {
+      blas.MatMul(M, N, K, X, W, Y);
+    }
     if (B == NULL) {
       return;
     }
@@ -37,7 +56,12 @@ class FCFunctor<platform::CPUDeviceContext, T> {
               .At(N);
       for (int i = 0; i < M; i++) {
         T* dst = Y + i * N;
-        compute(B, dst, dst, N);
+        T* src = nullptr;
+        if (weight_pass)
+          src = Y1_data + i * (N + 4);
+        else
+          src = dst;
+        compute(B, src, dst, N);
       }
     } else {
       auto compute =
@@ -48,7 +72,12 @@ class FCFunctor<platform::CPUDeviceContext, T> {
 #endif
       for (int i = 0; i < M; i++) {
         T* dst = Y + i * N;
-        compute(B, dst, dst, N);
+        T* src = nullptr;
+        if (weight_pass)
+          src = Y1_data + i * (N + 4);
+        else
+          src = dst;
+        compute(B, src, dst, N);
       }
     }
   }