… paddle/fluid/inference/tests/api/analyzer_bert_tester.cc
… paddle/fluid/operators/mkldnn/fc_mkldnn_op.cc
… python/paddle/fluid/tests/unittests/test_fc_op.py